枚举工具类优化，使用long替代int，避免枚举实例超过32个之后尴尬，基本应该不会超过64个

+++ easy-onlyoffice-sdk/src/main/java/com/github/onlyoffice/sdk/utils/EnumUtil.java
@@ -20,8 +20,8 @@ public class EnumUtil {
         if (one == null) {
             return false;
         }
-        int oneBit = 1 << one.ordinal();
-        int aggregateBits = Stream.of(multiple)
+        long oneBit = 1 << one.ordinal();
+        long aggregateBits = Stream.of(multiple)
                 .map(Enum::ordinal)
                 .reduce(0, (left, right) -> left | 1 << right);
